mypr/requirements.txt
Django>=4.0,<5.0
gunicorn>=20.0.4
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.0.0
